Kosmetyczne poprawki i porządki
Uporządkowałem trochę napisany przez Ciebie kod. Myślę, że poprawiło to czytelność, a i w projekcie jest trochę większy porządek. Daj znać co sądzisz o tych zmianach.
